# results/day9_async/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RegDocAI API",
    "description": "Day 9 API for synchronous review workflows plus PostgreSQL-backed asynchronous document jobs dispatched through Celery/Redis in production.",
    "version": "0.9.0"
  },
  "paths": {
...
    "/v1/documents/process": {
      "post": {
        "summary": "Process Document",
        "description": "Day 8-compatible synchronous endpoint retained for interactive review.",
        "operationId": "process_document_v1_documents_process_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_process_document_v1_documents_process_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs/process": {
      "post": {
        "summary": "Enqueue Document",
        "operationId": "enqueue_document_v1_jobs_process_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_enqueue_document_v1_jobs_process_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/batches/process": {
      "post": {
        "summary": "Enqueue Batch",
        "operationId": "enqueue_batch_v1_batches_process_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_enqueue_batch_v1_batches_process_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchSubmissionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs/{job_id}": {
      "get": {
        "summary": "Get Job",
        "operationId": "get_job_v1_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs/{job_id}/events": {
      "get": {
        "summary": "Get Job Events",
        "operationId": "get_job_events_v1_jobs__job_id__events_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobEventResponse"
                  },
                  "title": "Response Get Job Events V1 Jobs  Job Id  Events Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs/{job_id}/retry": {
      "post": {
        "summary": "Retry Job",
        "operationId": "retry_job_v1_jobs__job_id__retry_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/batches/{batch_id}": {
      "get": {
        "summary": "Get Batch",
        "operationId": "get_batch_v1_batches__batch_id__get",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/documents/{document_id}": {
      "get": {
        "summary": "Get Document",
        "operationId": "get_document_v1_documents__document_id__get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/documents/{document_id}/redactions/apply": {
      "post": {
        "summary": "Apply Redactions",
        "operationId": "apply_redactions_v1_documents__document_id__redactions_apply_post",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyRedactionsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplyRedactionsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/documents/{document_id}/artifacts/{artifact_name}": {
      "get": {
        "summary": "Download Artifact",
        "operationId": "download_artifact_v1_documents__document_id__artifacts__artifact_name__get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          },
          {
            "name": "artifact_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Artifact Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ApplyRedactionsRequest": {
        "properties": {
          "decisions": {
            "items": {
              "$ref": "#/components/schemas/RedactionDecision"
            },
            "type": "array",
            "title": "Decisions"
          }
        },
        "type": "object",
        "title": "ApplyRedactionsRequest"
      },
      "ApplyRedactionsResponse": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "redacted_count": {
            "type": "integer",
            "title": "Redacted Count"
          },
          "review_count": {
            "type": "integer",
            "title": "Review Count"
          },
          "retained_count": {
            "type": "integer",
            "title": "Retained Count"
          },
          "ignored_count": {
            "type": "integer",
            "title": "Ignored Count"
          },
          "artifact": {
            "$ref": "#/components/schemas/ArtifactLink"
          },
          "audit_artifact": {
            "$ref": "#/components/schemas/ArtifactLink"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "redacted_count",
          "review_count",
          "retained_count",
          "ignored_count",
          "artifact",
          "audit_artifact"
        ],
        "title": "ApplyRedactionsResponse"
      },
...
      "BatchResponse": {
        "properties": {
          "batch_id": {
            "type": "string",
            "title": "Batch Id"
          },
          "job_count": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Job Count"
          },
          "completed_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Completed Count"
          },
          "progress": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Progress"
          },
          "status_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Status Counts"
          },
          "jobs": {
            "items": {
              "$ref": "#/components/schemas/JobResponse"
            },
            "type": "array",
            "title": "Jobs"
          }
        },
        "type": "object",
        "required": [
          "batch_id",
          "job_count",
          "completed_count",
          "progress",
          "status_counts",
          "jobs"
        ],
        "title": "BatchResponse"
      },
      "BatchSubmissionResponse": {
        "properties": {
          "batch_id": {
            "type": "string",
            "title": "Batch Id"
          },
          "jobs": {
            "items": {
              "$ref": "#/components/schemas/JobResponse"
            },
            "type": "array",
            "title": "Jobs"
          }
        },
        "type": "object",
        "required": [
          "batch_id",
          "jobs"
        ],
        "title": "BatchSubmissionResponse"
      },
      "Body_enqueue_batch_v1_batches_process_post": {
        "properties": {
          "files": {
            "items": {
              "type": "string",
              "format": "binary"
            },
            "type": "array",
            "title": "Files"
          }
        },
        "type": "object",
        "required": [
          "files"
        ],
        "title": "Body_enqueue_batch_v1_batches_process_post"
      },
      "Body_enqueue_document_v1_jobs_process_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_enqueue_document_v1_jobs_process_post"
      },
      "Body_process_document_v1_documents_process_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_process_document_v1_documents_process_post"
      },
...
      "CheckboxResult": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence"
          },
          "bounding_box": {
            "$ref": "#/components/schemas/BoundingBox"
          }
        },
        "type": "object",
        "required": [
          "name",
          "state",
          "page",
          "confidence",
          "bounding_box"
        ],
        "title": "CheckboxResult"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobEventResponse": {
        "properties": {
          "sequence": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Sequence"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "progress": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Progress"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "sequence",
          "status",
          "stage",
          "progress",
          "message",
          "created_at"
        ],
        "title": "JobEventResponse"
      },
      "JobResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "batch_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch Id"
          },
          "source_filename": {
            "type": "string",
            "title": "Source Filename"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "progress": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Progress"
          },
          "current_stage": {
            "type": "string",
            "title": "Current Stage"
          },
          "attempt_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Attempt Count"
          },
          "max_retries": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Max Retries"
          },
          "celery_task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Celery Task Id"
          },
          "queue_mode": {
            "type": "string",
            "title": "Queue Mode"
          },
          "error_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Type"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "queued_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Queued At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "processing_seconds": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Processing Seconds"
          },
          "result_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Path"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "document_id",
          "source_filename",
          "status",
          "progress",
          "current_stage",
          "attempt_count",
          "max_retries",
          "queue_mode",
          "created_at"
        ],
        "title": "JobResponse"
      },
      "PageClassification": {
        "properties": {
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence"
          },
          "decision_source": {
            "type": "string",
            "title": "Decision Source"
          },
          "rule_reason": {
            "type": "string",
            "title": "Rule Reason"
          },
          "table_line_score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Table Line Score"
          }
        },
        "type": "object",
        "required": [
          "page",
          "label",
          "confidence",
          "decision_source",
          "rule_reason",
          "table_line_score"
        ],
        "title": "PageClassification"
      },
      "ProcessingResponse": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "source_filename": {
            "type": "string",
            "title": "Source Filename"
          },
          "status": {
            "$ref": "#/components/schemas/ProcessingStatus"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "processing_seconds": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Processing Seconds"
          },
          "page_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Page Count"
          },
          "classifications": {
            "items": {
              "$ref": "#/components/schemas/PageClassification"
            },
            "type": "array",
            "title": "Classifications"
          },
          "fields": {
            "items": {
              "$ref": "#/components/schemas/ExtractedField"
            },
            "type": "array",
            "title": "Fields"
          },
          "checkboxes": {
            "items": {
              "$ref": "#/components/schemas/CheckboxResult"
            },
            "type": "array",
            "title": "Checkboxes"
          },
          "tables": {
            "items": {
              "$ref": "#/components/schemas/TableArtifact"
            },
            "type": "array",
            "title": "Tables"
          },
          "redaction_candidates": {
            "items": {
              "$ref": "#/components/schemas/RedactionCandidate"
            },
            "type": "array",
            "title": "Redaction Candidates"
          },
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/ArtifactLink"
            },
            "type": "array",
            "title": "Artifacts"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "source_filename",
          "status",
          "created_at",
          "page_count"
        ],
        "title": "ProcessingResponse"
      },
      "ProcessingStatus": {
        "type": "string",
        "enum": [
          "UPLOADED",
          "QUEUED",
          "PREPROCESSING",
          "OCR_RUNNING",
          "TABLE_EXTRACTION",
          "REDACTION",
          "VALIDATION",
          "COMPLETED",
          "FAILED",
          "NEEDS_REVIEW"
        ],
        "title": "ProcessingStatus"
      },
...
    }
  }
}